Streamline handling of variables and tensors
The entire logic that was present to determine whether a given tensor is
created or loaded and whether when it will be set to zero and when not
is not also applicable to variables.
This is in order to prevent a bug in which result variables would be
loaded but not set to zero before their intended value is added to it.